+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 * Add ALSA asynchronous playback
 * Program can download and load some common missing libraries to run games
 * Remove multiple markers (#596)
+* Add more lua functions (savestates, ff)
 
 ### Changed